+++ app/src/main/java/org/mozilla/fenix/settings/quicksettings/QuickSettingsSheetDialogFragment.kt
@@ -11,24 +11,44 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
+import mozilla.components.feature.sitepermissions.SitePermissions
 import org.mozilla.fenix.R
+import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.mvi.ActionBusFactory
 import org.mozilla.fenix.mvi.getAutoDisposeObservable
 import org.mozilla.fenix.mvi.getManagedEmitter
 import org.mozilla.fenix.settings.PhoneFeature
+import kotlin.coroutines.CoroutineContext
 
 private const val KEY_URL = "KEY_URL"
 private const val KEY_IS_SECURED = "KEY_IS_SECURED"
-private const val KEY_IS_SITE_IN_EXCEPTION_LIST = "KEY_IS_SITE_IN_EXCEPTION_LIST"
+private const val KEY_SITE_PERMISSIONS = "KEY_SITE_PERMISSIONS"
 private const val REQUEST_CODE_QUICK_SETTINGS_PERMISSIONS = 4
 
 @SuppressWarnings("TooManyFunctions")
-class QuickSettingsSheetDialogFragment : BottomSheetDialogFragment() {
+class QuickSettingsSheetDialogFragment : BottomSheetDialogFragment(), CoroutineScope {
     private val safeArguments get() = requireNotNull(arguments)
     private val url: String by lazy { safeArguments.getString(KEY_URL) }
     private val isSecured: Boolean by lazy { safeArguments.getBoolean(KEY_IS_SECURED) }
-    private val isSiteInExceptionList: Boolean by lazy { safeArguments.getBoolean(KEY_IS_SITE_IN_EXCEPTION_LIST) }
     private lateinit var quickSettingsComponent: QuickSettingsComponent
+    private lateinit var job: Job
+
+    var sitePermissions: SitePermissions?
+        get() = safeArguments.getParcelable(KEY_SITE_PERMISSIONS)
+        set(value) {
+            safeArguments.putParcelable(KEY_SITE_PERMISSIONS, value)
+        }
+
+    override val coroutineContext: CoroutineContext get() = Dispatchers.IO + job
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        job = Job()
+    }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_quick_settings_dialog_sheet, container, false)
@@ -39,7 +59,7 @@ class QuickSettingsSheetDialogFragment : BottomSheetDialogFragment() {
         quickSettingsComponent = QuickSettingsComponent(
             rootView as ConstraintLayout, ActionBusFactory.get(this),
             QuickSettingsState(
-                QuickSettingsState.Mode.Normal(url, isSecured, isSiteInExceptionList)
+                QuickSettingsState.Mode.Normal(url, isSecured, sitePermissions)
             )
         )
     }
@@ -50,7 +70,7 @@ class QuickSettingsSheetDialogFragment : BottomSheetDialogFragment() {
         fun newInstance(
             url: String,
             isSecured: Boolean,
-            isSiteInExceptionList: Boolean
+            sitePermissions: SitePermissions?
         ): QuickSettingsSheetDialogFragment {
 
             val fragment = QuickSettingsSheetDialogFragment()
@@ -59,7 +79,7 @@ class QuickSettingsSheetDialogFragment : BottomSheetDialogFragment() {
             with(arguments) {
                 putString(KEY_URL, url)
                 putBoolean(KEY_IS_SECURED, isSecured)
-                putBoolean(KEY_IS_SITE_IN_EXCEPTION_LIST, isSiteInExceptionList)
+                putParcelable(KEY_SITE_PERMISSIONS, sitePermissions)
             }
             fragment.arguments = arguments
             return fragment
@@ -70,8 +90,13 @@ class QuickSettingsSheetDialogFragment : BottomSheetDialogFragment() {
         if (arePermissionsGranted(requestCode, grantResults)) {
             val feature = requireNotNull(PhoneFeature.findFeatureBy(permissions))
             getManagedEmitter<QuickSettingsChange>()
-                .onNext(QuickSettingsChange.PermissionGranted(feature))
+                .onNext(QuickSettingsChange.PermissionGranted(feature, sitePermissions))
         }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        job.cancel()
     }
 
     private fun arePermissionsGranted(requestCode: Int, grantResults: IntArray) =
@@ -85,13 +110,30 @@ class QuickSettingsSheetDialogFragment : BottomSheetDialogFragment() {
                     is QuickSettingsAction.SelectBlockedByAndroid -> {
                         requestPermissions(it.permissions, REQUEST_CODE_QUICK_SETTINGS_PERMISSIONS)
                     }
-                    is QuickSettingsAction.DismissDialog -> dismiss()
+                    is QuickSettingsAction.TogglePermission -> {
+
+                        launch {
+                            sitePermissions = quickSettingsComponent.toggleSitePermission(
+                                context = requireContext(),
+                                featurePhone = it.featurePhone,
+                                url = url,
+                                sitePermissions = sitePermissions
+                            )
+
+                            launch(Dispatchers.Main) {
+                                getManagedEmitter<QuickSettingsChange>()
+                                    .onNext(QuickSettingsChange.Stored(it.featurePhone, sitePermissions))
+
+                                requireContext().components.useCases.sessionUseCases.reload.invoke()
+                            }
+                        }
+                    }
                 }
             }
 
         if (isVisible) {
             getManagedEmitter<QuickSettingsChange>()
-                .onNext(QuickSettingsChange.PromptRestarted)
+                .onNext(QuickSettingsChange.PromptRestarted(sitePermissions))
         }
     }
 }